+++ src/app/heroes/views/heroe/heroe.component.html
@@ -2,7 +2,7 @@
   <h1>{{ "HEROE.TITLE_HEROE" | transloco }}</h1>
   <hr />
   <div class="example-button-row">
-    <button routerLink="/heroes/heroes" mat-raised-button color="primary">
+    <button routerLink="/heroes" mat-raised-button color="primary">
       {{ "HEROE.RETURN" | transloco }}
     </button>
   </div>
