+++ src/app/heroes/views/heroes/heroes-routing.module.ts
@@ -1,0 +1,18 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+
+import { HeroesComponent } from './heroes.component';
+
+const route: Routes = [
+  {
+    path: '',
+    component: HeroesComponent,
+  },
+];
+
+@NgModule({
+  declarations: [],
+  imports: [RouterModule.forChild(route)],
+  exports: [RouterModule],
+})
+export class HeroesRoutingModule {}
