+++ src/app/home/home-routing.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { HomeComponent } from './home.component';
+
+
+
+const route: Routes = [
+  {
+    path: '',
+    component: HomeComponent,
+  },
+];
+
+@NgModule({
+  declarations: [],
+  imports: [RouterModule.forChild(route)],
+  exports: [RouterModule],
+})
+
+export class HomeRoutingModule { }
