+++ src/app/shared/navbar/navbar.component.html
@@ -7,7 +7,7 @@
       <button mat-button routerLink="/home">
         {{ "NAVBAR.HOME" | transloco }}
       </button>
-      <button mat-button routerLink="/heroes/heroes">
+      <button mat-button routerLink="/heroes">
         {{ "HEROES.HEROES" | transloco }}
       </button>
     </div>
